refactor: formatDate 함수 util 함수로 분리
