+++ src/app/utils/formatDate.ts
@@ -1,0 +1,11 @@
+// 날짜 표시 변경 함수
+const formatDate = (dateString: string) => {
+  const date = new Date(dateString);
+  return new Intl.DateTimeFormat('ko-KR', {
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric',
+  }).format(date);
+};
+
+export default formatDate;
